+++ pqfmlib/maps/xyz.py
@@ -478,6 +478,7 @@ class XYZProjectiveQFM(BaseProjectiveQFM):
             bool(self.simulation),
             bool(self.fakebackend),
             bool(self.ideal),
+            self.expectation_method,
             int(self.shots),
             str(self.ibm_qpu),
             int(self.q_enc),
@@ -625,6 +626,7 @@ class XYZProjectiveQFM(BaseProjectiveQFM):
             self.backend,
             self.estimator,
             simulation=True,
+            expectation_method=self.expectation_method,
             shots=self.shots,
             base_folder=None,
         )
@@ -825,6 +827,7 @@ class XYZProjectiveQFM(BaseProjectiveQFM):
             simulation=self.simulation,
             fakebackend=self.fakebackend,
             resource_estimation=self.resource_estimation,
+            expectation_method=self.expectation_method,
             shots=self.shots,
             base_folder=self.base_folder,
             fixed_circuit=self.fixed_circuit_file_name,